Restructured the interface. There is now a separate BPMclass file to keep it tidy and separated. I added methods to it to match how other nodes work.
I also made a PV class so that all the PVs use the same methods within interface. The PVs are defined by assigning a node and the method(s) within the node the PV needs to interact with. It also keeps track of meta information about the PV, node, and method. This cleaned up a lot of code and removes needing to parse strings to determine how interface should function.

A lot of bug were also fixed in the process.