Unified string keys for PyORBIT class identification across scripts. This allows for get_type to be used in the model to easily determine element types for the config file.
I added the phase_offset file as a separate argument with a default of 0. The cavity and BPM virtual devices now no longer give a random phase offset if they don't receive one. A randomized offset will need to be created in virtual_accelerator.py if the user desires something random each time.