I added functionality for quad doublets. This required rewriting how specific devices use get_settings. I was doing it poorly anyways, so this was a needed update. Now, each device handles get_setting separately and uses their parent get_setting function. Which I believe was the original intention.
I also changed the server to say that each device has a list of models attached to it. If only a model name is given, server turns it into a single item list. This allows for looping over model names that are attached to a singular device, like the doublet.