A few updates in this commit. First, a dump bunch node was created in pyorbit_child_nodes.py. The file name is a PV defined by the bunch dumper device in virtual_devices_SNS.py. This allows a node to be added that saves the bunch at the location in the virtual accelerator with a name that can be modified using EPICS.
I've also modified sns_linac.xml and pyorbit_lattice_factory.py. First, the previous sequence of each sequence is now an argument in the xml file and the lattice factory uses that to stitch together sequences. This allows for the removal of the complicated if statements in virtual_accelerator.py. Second, pyorbit_lattice_factory.py now adds the BPM and WireScanner nodes directly to the lattice. They are no longer child nodes of Marker nodes and don't need to be added by virtual_accelerator.py directly.