Fixed bug where changing the beam current didn't affect space charge calculations. Setting the beam current now sets the macroparticle size for the sns linac. THIS HAS NOT BEEN ADDED TO BTF OR IDUMP!
Get values for VirtualAccelerator can now accept a list to return only the desired values in a dictionary.

Added physics nodes and devices that calculate Twiss parameters at the location and publish the PVs. All PVs use 'Physics' in their name. Also added the maximum drift length as an argument that can be set to control the frequency of physics and space charge nodes.